Subset.locationOf fix
The `Subset.locationOf` function must delegate to the decorated configuration and complete the given key in order to be compatible with all other functions within it. Currently it can find a key but not report its location. This change fixes that.

PS: I had to fix the include pattern in the Gradle build file because the current pattern does not lead to any matches.